requirements.txt
openai
python-dotenv
jira
PyPDF2
python-docx